requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-dotenv